refactor(group-quiz): extract action layer into useGroupQuizActions — logic 100%
Move all 9 room actions (start/cancel/leave/submitAnswers/joinHere/
backToLobby/restart/synthesize/backToDashboard) verbatim into
_hooks/use-group-quiz-actions.ts, and the shared answer helpers
(formatAnswerForStorage/hasAnswer/LocalAnswers) into _lib/answers.ts so
the page render and the hook share them without a cycle. Optimistic
setSession patches preserved exactly. Pruned now-unused groupQuizService/
isOverloadedError/QuestionValue imports.

group-quiz/[code]/page is now logic-complete: every state/effect/handler
lives in useGroupQuizSession/useGroupQuizShare/useGroupQuizActions; only
the render + orchestrator wiring remain. Behavior-identical;
page.tsx 1182->996 LOC (1637->996, -39% across the sweep).
